Initial Commit
A second implementation of Dimtry Vyukov's [non-intrusive mpsc fifo](http://www.1024cores.net/home/lock-free-algorithms/queues/non-intrusive-mpsc-node-based-queue).
This first implementation I did is [here](https://github.com/winksaville/test-mpscfifo)
this one differs by having a Cell_t and allocating the stub in the fifo
and having another one in the Msg_t. So as with my previous implementation
it actually is intrusive but that doesn't bother me.

But what I don't like it that it's 30% to 50% slower see the README.md
for some details.